added some validation to notification dismiss function. Added tests

+++ src/models/user/user.notifications/user.notifications.methods.ts
@@ -26,6 +26,15 @@ export async function dismissNotification(
   this: IUserDocument,
   notificationId: string,
 ): Promise<INotificationDocument[]> {
+  const notificationDocument = await NotificationModel.findById(notificationId);
+
+  if (!notificationDocument) {
+    throw new Error("Unable to find notification by id");
+  }
+  if (notificationDocument.targetId !== this.id) {
+    throw new Error("Dismiss notifications: illegal operation - targetId and userId don't match");
+  }
+
   await NotificationModel.findByIdAndDelete(notificationId);
   this.notifications = this.notifications.filter((notification) =>
     notification !== notificationId,
